+++ common/src/main/kotlin/com/cobblemon/mod/common/command/ClearPCCommand.kt
@@ -10,7 +10,7 @@ package com.cobblemon.mod.common.command
 
 import com.cobblemon.mod.common.api.permission.CobblemonPermissions
 import com.cobblemon.mod.common.util.commandLang
-import com.cobblemon.mod.common.util.party
+import com.cobblemon.mod.common.util.pc
 import com.cobblemon.mod.common.util.permission
 import com.mojang.brigadier.Command
 import com.mojang.brigadier.CommandDispatcher
@@ -19,33 +19,22 @@ import net.minecraft.command.argument.EntityArgumentType
 import net.minecraft.server.command.CommandManager
 import net.minecraft.server.command.ServerCommandSource
 
-object ClearPartyCommand {
+object ClearPCCommand {
 
-    private const val NAME = "clearparty"
+    private const val NAME = "clearpc"
     private const val PLAYER = "player"
 
     fun register(dispatcher : CommandDispatcher<ServerCommandSource>) {
         val command = CommandManager.literal(NAME)
-             .permission(CobblemonPermissions.CLEAR_PARTY)
-             .then( CommandManager.argument(PLAYER, EntityArgumentType.players()).executes(::execute)
-                )
+            .permission(CobblemonPermissions.CLEAR_PC)
+            .then(CommandManager.argument(PLAYER, EntityArgumentType.players()).executes(::execute))
         dispatcher.register(command)
     }
 
     private fun execute(context: CommandContext<ServerCommandSource>) : Int {
         val target = EntityArgumentType.getPlayer(context, "player")
-        val party = target.party()
-
-        val pokemonList = party.filterNotNull()
-        if (pokemonList.isEmpty()) {
-            context.source.sendError(commandLang("$NAME.nonethere", target.displayName))
-            return 0
-        }
-
-        for (pokemon in pokemonList) {
-            party.remove(pokemon)
-        }
-
+        val pc = target.pc()
+        pc.toList().forEach(pc::remove)
         context.source.sendFeedback({ commandLang("$NAME.cleared", target.displayName) }, true)
         return Command.SINGLE_SUCCESS
     }